python-services/requirements.txt
fastapi
uvicorn
gtts
python-dotenv